backend/pyproject.toml
[project]
name = "homelan-backend"
version = "0.1.0"
description = "HomeLan - Network Dashboard Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "websockets>=13.0",
    "httpx>=0.28.0",
    "netifaces>=0.11.0",
    "paramiko>=3.4.0",
    "dnspython>=2.6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
